Problem: Using pg_dump/restore needs storage and slow
Solution: Use pgcopydb to migrate extension config table and historical
data.

Previously we were using pg_dump in parallel & pg_restore in serial mode
due issues with timescale while restoring extension data in parallel. In
addition to that pg_dump needs storage to dump the contents which is
not optimal because the user has to predecide on the storage volume
while initiating the migration.

Pgcopydb based solution is better optimized for parallel data copy and
it just streams table data from source to target.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>