Run timescaledb_post_restore() after schema migration
When timescaledb_post_restore() is run, it will attempt to start
background jobs. If retention jobs exist on hypertables, attempting to
start them will trigger a code path in TimescaleDB which validates that
the hypertable which the TimescaleDB catalog references exists. If only
the extension configuration is present, but not the schema for tables,
this check will fail, and abort the migration.

This change moves the call to timescaledb_post_restore() to after the
post-data section, avoiding the above error.

Note: Initially we moved the call to timescaledb_post_restore() to after
the pre-data section, but this causes problems with foreign key and index
creation.