Problem: Privileges/grants are not migrated
We use the --no-owner and --no-privileges flags during the PG to TS dump, and the --no-acl and --no-owner flags for TS to TS, which omits GRANT statements, leading to missing privileges on the target.

This is not specific to live-migration, we recommend the same flags in [other migration methods](https://arc.net/l/quote/ofbczfsm) as well.

Solution: Dump pre-data without --no-owner and --no-privileges flag,
extract GRANT, REVOKE, ALTER.. OWNER .. statments and apply on target.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>