Problem: Replication fails on tables with generated column (#90)
Solution: Skip generated columns while decoding logical messages. The transform process prepares an in-memory cache to hold all the tables with generated columns during the initialization of the transform. Generally, we don't expect to have many tables with generated columns, so keeping them in memory is a better trade-off and avoids performance bottlenecks.

Other possible solutions: Replacing the generated value with DEFAULT. This was not feasible because there is no way to mark keywords in the EXECUTE JSON array in the .sql file.

Related PR which skips generated column for table-data copy: dimitri/pgcopydb#517

Fixes timescale/team-data-onboarding#57

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>