Fix resume on interruped index creation
If we interrupt the process while creating an index, the summary catalog
will show the index command as `CREATE INDEX ..`. When resuming,
pgcopydb alters the command to `CREATE INDEX IF NOT EXISTS ..` to avoid errors if index already created on target. However, summary_lookup_index() will change it back to `CREATE INDEX ..` because it replaces the command from the summary catalog table.

Solution: Change the index definition just before executing it against the target.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>