(hotfix) Skip DML on matview (#105)
This commit renames all materialized views on target instances to
`temp_{matview}` and creates a view with the same name as matview but
pointing to the renamed matview. It also installs trigger on the matview
to ignore all incoming DML operations. During the live replication
stops, the reverse operation performed(i.e. renamed materialized view
will be renamed back to it's original name).

## Alternatives considered

- Filter DML operations for matview during CDC
- Adopt --exclude-table-data flag for CDC and let the customer to
provide the list of matviews to filter

Both of the above can be potential solutions for future, but it needs
changes to pgcopydb and might take considerable time to implement.

The current solution can be considered as a hotfix/quick fix to unblock
the customer with minimal effort.

Reported by https://github.com/timescale/Support-Dev-Collab/issues/1712

Fixes https://github.com/timescale/team-data-onboarding/issues/153

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>