Problem: Vacuum worker exits early after table copy (#116)
Vacuum workers are signalled to STOP as soon as the table copy workers
are done. This is incorrect as the index process can still process
indexes even after the termination of copy workers and signal vacuum
worker once all indexes are done.

Solution: Create an index supervisor to wait for index workers to
complete and send STOP to vacuum. This preserves the order of worker
termination.

Upstream PR: https://github.com/dimitri/pgcopydb/pull/815

(NOTE: The upstream already had supervisors for both index and vacuum
which was added as part of the large PR. We are only interested in
adding index supervisor only here)

---------

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>