Fail fast when initial data copy fails (#143)
By default, pgcopydb waits for all processes to exit before reporting
errors. Sometimes, this might give customers the wrong impression that
everything is going well and they leave the system running for a long
period. In reality, something may have failed and will only be reported
at the end of the copy. IMHO, this is not a good user experience.
Fortunately, pgcopydb has a flag named --fail-fast which helps to report
errors as soon as they occur.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>