Fix data loss issue during partitioned table resume in pgcopydb
We first TRUNCATE the table before queuing copy operations for pgcopydb partitioned tables. While this behavior works under normal circumstances, it can lead to data loss when resuming after an interruption. This happens because TRUNCATE removes already copied contents, and the tool skips copying table parts that are marked as already copied.

This fix ensures that TRUNCATE is executed only if none of the table parts have been copied.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>